+++ BelibaHoma.BLL/Models/TutorModel.cs
@@ -54,10 +54,10 @@ namespace BelibaHoma.BLL.Models
         public string PhoneNumber { get; set; }
 
         [Display(Name = "שעות חברותא (השנה)")]
-        public int TutorHoursBonding { get; set; }
+        public double TutorHoursBonding { get; set; }
 
         [Display(Name = "שעות חניכה (השנה)")]
-        public int TutorHours { get; set; }
+        public double TutorHours { get; set; }
 
 
         [Display(Name = "רמת חניכה בפיזיקה")]
